Refactor translateWhereToQuery
