+++ utils/OData.js
@@ -6,7 +6,6 @@ const config = require('../conf/default')
 const utils = require('./utils')
 const modelUtils = require('./modelUtils')
 const requestUtils = require('./requestUtils')
-const translateUtils = require('./translateWhereToQuery')
 
 var o
 module.exports = (sdk) => {
@@ -559,7 +558,7 @@ OData.prototype._where = function (where, inst) {
   }
 
   var filter = Object.keys(where).reduce((str, key) => {
-    return translateUtils.translateWhereToQuery(where, str, key)
+    return utils.translateWhereToQuery(where, str, key)
   }, '')
 
   if (filter !== '') {
